Refactor @cucumber/query - combine into a single Query class
