+++ react/javascript/Makefile
@@ -12,5 +12,11 @@ testdata/%.ndjson: testdata/%.feature
 $(ALL_NDJSON_FILE): $(FEATURE_FILES)
 	./node_modules/@cucumber/fake-cucumber/bin/fake-cucumber --predictable-ids --format ndjson $^ > $@
 
+.tested: .tested-storybook
+
+.tested-storybook:
+	./node_modules/.bin/build-storybook
+	touch $@
+
 clean:
 	rm -f $(GHERKIN_DOCUMENT_NDJSON_FILES) $(ALL_NDJSON_FILE) $(ALL_PROTOBUF_FILE)
