+++ react/javascript/src/CucumberQueryContext.ts
@@ -1,0 +1,4 @@
+import React from 'react'
+import CucumberQuery from '@cucumber/query'
+
+export default React.createContext(new CucumberQuery())
