--- react/javascript/src/StepMatchArgumentsQueryContext.ts
@@ -1,4 +1,0 @@
-import React from 'react'
-import { StepMatchArgumentsQuery } from '@cucumber/query'
-
-export default React.createContext(new StepMatchArgumentsQuery())
