--- react/javascript/src/TestResultsQueryContext.ts
@@ -1,4 +1,0 @@
-import React from 'react'
-import { TestResultsQuery } from '@cucumber/query'
-
-export default React.createContext(new TestResultsQuery())
