+++ react/javascript/src/components/app/Wrapper.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import { messages } from '@cucumber/messages'
 import { GherkinQuery } from '@cucumber/gherkin'
-import { StepMatchArgumentsQuery, TestResultsQuery } from '@cucumber/query'
+import CucumberQuery from '@cucumber/query'
 import GherkinQueryContext from '../../GherkinQueryContext'
-import StepMatchArgumentsQueryContext from '../../StepMatchArgumentsQueryContext'
-import TestResultsQueryContext from '../../TestResultsQueryContext'
+import CucumberQueryContext from '../../CucumberQueryContext'
 
 interface IProps {
   envelopes: messages.IEnvelope[]
@@ -12,23 +11,19 @@ interface IProps {
 
 const Wrapper: React.FunctionComponent<IProps> = ({ envelopes, children }) => {
   const gherkinQuery = new GherkinQuery()
-  const testResultsQuery = new TestResultsQuery()
-  const stepMatchArgumentsQuery = new StepMatchArgumentsQuery()
+  const cucumberQuery = new CucumberQuery()
   envelopes.forEach(envelope => {
     gherkinQuery.update(envelope)
-    testResultsQuery.update(envelope)
-    stepMatchArgumentsQuery.update(envelope)
+    cucumberQuery.update(envelope)
   })
 
   return (
     <div className="cucumber-react">
-      <StepMatchArgumentsQueryContext.Provider value={stepMatchArgumentsQuery}>
-        <TestResultsQueryContext.Provider value={testResultsQuery}>
-          <GherkinQueryContext.Provider value={gherkinQuery}>
-            {children}
-          </GherkinQueryContext.Provider>
-        </TestResultsQueryContext.Provider>
-      </StepMatchArgumentsQueryContext.Provider>
+      <CucumberQueryContext.Provider value={cucumberQuery}>
+        <GherkinQueryContext.Provider value={gherkinQuery}>
+          {children}
+        </GherkinQueryContext.Provider>
+      </CucumberQueryContext.Provider>
     </div>
   )
 }
