+++ react/javascript/src/components/gherkin/Step.tsx
@@ -3,10 +3,9 @@ import DataTable from './DataTable'
 import Keyword from './Keyword'
 import DocString from './DocString'
 import { messages } from '@cucumber/messages'
-import TestResultQueryContext from '../../TestResultsQueryContext'
+import CucumberQueryContext from '../../CucumberQueryContext'
 import UriContext from '../../UriContext'
 import GherkinQueryContext from '../../GherkinQueryContext'
-import StepMatchArgumentsQueryContext from '../../StepMatchArgumentsQueryContext'
 import ErrorMessage from './ErrorMessage'
 import StepContainer from './StepContainer'
 import Attachment from './Attachment'
@@ -23,16 +22,13 @@ const Step: React.FunctionComponent<IProps> = ({
   renderMessage,
 }) => {
   const gherkinQuery = React.useContext(GherkinQueryContext)
-  const testResultQuery = React.useContext(TestResultQueryContext)
-  const stepMatchArgumentsQuery = React.useContext(
-    StepMatchArgumentsQueryContext
-  )
+  const cucumberQuery = React.useContext(CucumberQueryContext)
   const uri = React.useContext(UriContext)
 
   const pickleStepIds = gherkinQuery.getPickleStepIds(uri, step.location.line)
-  const pickleStepResults = testResultQuery.getPickleStepResults(pickleStepIds)
-  const testResult = testResultQuery.getWorstResult(pickleStepResults)
-  const attachments = testResultQuery.getPickleStepAttachments(pickleStepIds)
+  const pickleStepResults = cucumberQuery.getPickleStepResults(pickleStepIds)
+  const testResult = cucumberQuery.getWorstResult(pickleStepResults)
+  const attachments = cucumberQuery.getPickleStepAttachments(pickleStepIds)
 
   const stepTextElements: JSX.Element[] = []
 
@@ -44,7 +40,7 @@ const Step: React.FunctionComponent<IProps> = ({
           // the background step will be rendered as undefined (even if there are matching step definitions). This
           // is not ideal, but it is rare enough that we don't care about it for now.
           []
-        : stepMatchArgumentsQuery.getStepMatchArgumentsLists(pickleStepIds[0])
+        : cucumberQuery.getStepMatchArgumentsLists(pickleStepIds[0])
     if (stepMatchArgumentsLists.length === 1) {
       // Step is defined
       const stepMatchArguments = stepMatchArgumentsLists[0].stepMatchArguments
